requirements.txt
Django==4.0.1
requests-html~=0.10.0